Module#class_attribute avoid redefining the setter
Fix: https://github.com/rails/rails/issues/53606

This allow to be able to prepend a module that override the setter
method, and for it to work even on subclasses.

This still doesn't work for getter methods, and has never worked,
hence why I'm a bit on the fence about supporting this.

This keep the new implementation performance when assigning
on the class where the attribute is defined, but revert to
the old implementation performance when assigning on a child
class.