Implement `#silenced?` protocol for ExplainSubscriber
It's a noop the vast majority of the time, might as well filter it
out.

Right now that doesn't help a lot because the `sql.active_record`
event always have at least one active subscriber for `RuntimeRegistry`,
but if we refactor that to no longer depend on a subscriber, we
could potentially shortcircuit all of ActiveSupport::Notification
in the happy path.